requirements.txt
celery==5.6.0
Django==6.0
django_celery_results==2.6.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
redis==7.1.0
requests==2.32.5
